Refactor animate.c to use floating point seconds.
- Packet extraction in sever.c converts animation duration from milliseconds to float seconds.
- Updated animation function parameters to receive delta time as float seconds.
- Changed timing variables for Animate and object_node structs from milliseconds to seconds.
- Added conversion to milliseconds for RoomAnimateSingle() and MoveSingle() to work with legacy systems.

+++ clientd3d/dialog.c
@@ -506,9 +506,9 @@ INT_PTR CALLBACK DescDialogProc(HWND hDlg, UINT message, WPARAM wParam, LPARAM l
 /************************************************************************/
 /* 
 * AnimateDescription:  Animate bitmap in description dialog.
-*   dt is number of milliseconds since last time animation timer went off.
+*   dt is number of seconds since last time animation timer went off.
 */
-void AnimateDescription(int dt)
+void AnimateDescription(float dt)
 {
 	int old_group;
 	DescDialogStruct *info;
